% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_distribution.R
\name{check_distribution}
\alias{check_distribution}
\title{Classify the distribution of a model-family using machine learning}
\usage{
check_distribution(model)
}
\arguments{
\item{model}{A model (that should response to \code{residuals()}).}
}
\description{
This function uses an internal random forest model to classify the distribution from a model-family.
}
\note{
This function is somewhat experimental and might be improved in future releases.
}
\examples{
library(lme4)
model <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
check_distribution(model)

}
