% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outlier.R
\name{check_outlier}
\alias{check_outlier}
\alias{check_outlier.default}
\title{Check for influential observations}
\usage{
check_outlier(x, ...)

\method{check_outlier}{default}(x, print_outliers = TRUE,
  drop_outliers = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Currently not used.}

\item{print_outliers}{Optional argument to list individual outliers. Default set as \code{TRUE}.}

\item{drop_outliers}{Optional argument to drop outliers and return a data frame minus outliers. Default set as \code{FALSE}.}
}
\value{
Check (message) on whether outliers were detected or not, as well as an optional
data frame minus outliers (i.e., \code{drop_outliers = TRUE}).
}
\description{
Checks for and locates influential observations (i.e., "outliers") via Cook's Distance.
}
\details{
Performs a Cook's distance test to check for influential observations. Those greater than 4/n, are considered
outliers. This relatively conservative threshold is useful only for detection, rather than justificaiton for automatic observation
deletion. If users opt to drop observations that may be problematic, they may do so by specifying \code{drop_outliers = TRUE}.
}
\examples{
mt1 <- mtcars[, c(1,3)] # select only mpg and disp (continuous)
oo <- data.frame(mpg = c(37, 40), disp = c(300, 400)) # create some fake outliers
mt2 <- rbind(mt1, oo) # attach outliers to main df
model <- lm(disp ~ mpg, data = mt2) # fit model with outliers
check_outlier(model, print_outliers = TRUE, drop_outliers = FALSE)

}
\references{
Cook, R. D. (1977). Detection of influential observation in linear regression. Technometrics, 19(1), 15-18.
}
