% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_priors.R
\name{check_priors}
\alias{check_priors}
\alias{check_priors.stanreg}
\title{Prior predictive checks}
\usage{
check_priors(model = NULL, ...)

\method{check_priors}{stanreg}(model = NULL, predictors = NULL, ...)
}
\arguments{
\item{model}{A Bayesian model of class \code{stanreg} or \code{brmsfit}. Note that
this model should include draws from the prior predictive distribution. This
can be achieved, e.g., by using \code{\link[bayestestR:unupdate]{bayestestR::unupdate()}} on the fitted
model, or setting \code{prior_PD = TRUE} (when using \strong{rstanarm}) or
\code{sample_prior = "only"} (when using \strong{brms}).}

\item{...}{Currently not used.}

\item{predictors}{Character vector with names of one or more model predictors
for which prior predictive checks should be visualized.}
}
\value{
A data frame of simulated responses and the original response vector.
}
\description{
Simulates from the prior marginal distribution of the data to assess the
consistency of the chosen priors with domain knowledge (\emph{Gabry et al. 2019})
and creates a visualization from the prior predictive checks.
}
\details{
Assessing priors based on the prior marginal distribution for the data
provides several methodological advantages:
\itemize{
\item It reflects the interplay between the prior distribution on the parameters
and the likelihood.
\item It is a vital component of understanding how prior distributions actually
work for a given problem.
\item It explicitly reflects the idea that we cannot fully understand the prior
by fixing all except one parameter and assessing the effect of the
unidimensional marginal prior.
\item Instead, we need to assess the effect of the prior as a multivariate
distribution.
\item The prior distribution over the data enables us to extend the concept of a
weakly informative prior to be more aware of the role of the
likelihood.
}

A prior leads to a weakly informative joint prior data-generating process if
draws from the prior data-generating distribution could represent any data
set that could plausibly be observed. Furthermore, there should be
no mass on completely implausible data sets. Generating simulated
data sets can be used to investigate the variability and multivariate
structure of the distribution.
}
\examples{
\dontrun{
model <- insight::download_model("stan_prior_checks_1")
plot(check_priors(model, "mmse"))
}
}
\references{
Gabry, J., Simpson, D., Vehtari, A., Betancourt, M., & Gelman, A. (2019).
Visualization in Bayesian Workflow. Journal of the Royal Statistical Society
Series A: Statistics in Society, 182(2), 389–402. \doi{10.1111/rssa.12378}
}
