% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_reliability.R
\name{check_reliability}
\alias{check_reliability}
\alias{check_reliability.estimate_grouplevel}
\title{Variability-Over-Uncertainty Ratio (d-vour) for Random Effects Reliability}
\usage{
check_reliability(x, ...)

\method{check_reliability}{estimate_grouplevel}(x, ...)
}
\arguments{
\item{x}{A model object (or from \code{\link[modelbased:estimate_grouplevel]{modelbased::estimate_grouplevel()}}).}

\item{...}{Currently not used.}

\item{n_trials}{to do...}
}
\description{
TODO: Add description.
}
\details{
TODO: Add details.

Intepretation: d-vour corresponds to: between-groups variability /  (between-groups variability + within-group variability)
A d-vour of 3/4 (0.75) means that there is 3 times more variability between groups than within groups (3/1),
and a d-vour of less than 0.5 means that there is more variability within groups than between groups (which is
bad if the goal is to analyze group-level effects).
}
\examples{
\dontshow{if (require("lme4") & require("glmmTMB")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df <- read.csv("https://raw.githubusercontent.com/easystats/circus/refs/heads/main/data/illusiongame.csv")

m <- lme4::lmer(RT ~ (1 | Participant), data = df)
check_reliability(m)
m <- glmmTMB::glmmTMB(RT ~ (1 | Participant), data = df)
check_reliability(m)

m <- lme4::lmer(RT ~ (1 | Participant) + (1 | Trial), data = df)
check_reliability(m)
m <- glmmTMB::glmmTMB(RT ~ (1 | Participant) + (1 | Trial), data = df)
check_reliability(m)

m <- lme4::lmer(RT ~ Illusion_Difference + (Illusion_Difference | Participant) + (1 | Trial), data = df)
check_reliability(m)
m <- glmmTMB::glmmTMB(RT ~ Illusion_Difference + (Illusion_Difference | Participant) + (1 | Trial), data = df)
check_reliability(m)
\dontshow{\}) # examplesIf}
}
\references{
TODO.
}
\seealso{
Other functions to check model assumptions and and assess model quality: 
\code{\link{check_autocorrelation}()},
\code{\link{check_collinearity}()},
\code{\link{check_convergence}()},
\code{\link{check_heteroscedasticity}()},
\code{\link{check_homogeneity}()},
\code{\link{check_model}()},
\code{\link{check_outliers}()},
\code{\link{check_overdispersion}()},
\code{\link{check_predictions}()},
\code{\link{check_singularity}()},
\code{\link{check_zeroinflation}()}
}
\concept{functions to check model assumptions and and assess model quality}
