% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_residuals.R
\name{check_residuals}
\alias{check_residuals}
\title{Check uniformity of simulated residuals}
\usage{
check_residuals(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=simulated_residuals]{simulated_residuals()}}.}

\item{...}{Passed down to \code{\link[DHARMa:testUniformity]{DHARMa::testUniformity()}}}
}
\description{
\code{check_residuals()} checks generalized linear (mixed) models for uniformity
of randomized quantile residuals, which can be used to identify typical model
misspecification problems, such as over/underdispersion, zero-inflation, and
residual spatial and temporal autocorrelation.
}
