% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_smooth.R
\name{check_smooth}
\alias{check_smooth}
\title{Check Smooth Term in GAMs}
\usage{
check_smooth(x, iterations = 400, ...)
}
\arguments{
\item{x}{An \link{mgcv} GAM model.}

\item{iterations}{Number of permutations.}

\item{...}{Other arguments to be passed to other functions (not used for now).}
}
\description{
Randomization test looking for evidence of residual pattern attributable to covariates of each smooth.
}
\examples{
if (require("mgcv")) {
  model <- mgcv::gam(Sepal.Length ~ s(Petal.Length, k = 3) + s(Petal.Width, k = 5), data = iris)
  check_smooth(model)
}
}
