% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_symmetry.R
\name{check_symmetry}
\alias{check_symmetry}
\title{Check distribution symmetry}
\usage{
check_symmetry(x, ...)
}
\arguments{
\item{x}{Model or numeric vector}

\item{...}{Not used.}
}
\description{
Uses Hotelling and Solomons test of symmetry by testing if the standardized
nonparametric skew (\eqn{\frac{(Mean - Median)}{SD}}) is different than 0.
\cr\cr
This is an underlying assumption of Wilcoxon signed-rank test.
}
\examples{
V <- suppressWarnings(wilcox.test(mtcars$mpg))
check_symmetry(V)

}
