% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-DataObject.R
\name{check_variable_type}
\alias{check_variable_type}
\title{Check Variable Type}
\usage{
check_variable_type(var)
}
\arguments{
\item{var}{A vector containing the data to be checked.}
}
\value{
A character string indicating the variable type. Possible
 values are "binary", "category", or "continuous".
}
\description{
This function takes a vector \code{var} as input and determines
 its variable type.
}
\details{
The function performs checks on non-NA values of the input
 vector \code{var}. If \code{var} is a character, or factor, it is
 considered a "category". If it contains only 0s and 1s, it is considered
 "binary". If it is numeric, it is considered "continuous".
}
\examples{
# Example 1: Binary variable
var1 <- c(0, 1, 1, 0, 1)
check_variable_type(var1)

# Example 2: Categorical variable
var2 <- factor(c("A", "B", "A", "C", "B"))
check_variable_type(var2)

# Example 3: Continuous variable
var3 <- c(10.5, 15.2, 12.8, 9.7, 13.4)
check_variable_type(var3)

}
