% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_distribution.R
\docType{data}
\name{classify_distribution}
\alias{classify_distribution}
\title{Classify the distribution of a model-family using machine learning}
\description{
Classify the distribution of a model-family using machine learning
}
\details{
The trained model to classify distributions, which is used by the
\code{check_distribution()} function.
}
\keyword{data}
