% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-backend.R
\name{continuous_x_binary}
\alias{continuous_x_binary}
\title{continuous_x_binary Function}
\usage{
continuous_x_binary(a, b)
}
\arguments{
\item{a}{A numeric vector representing the continuous variable.}

\item{b}{A logical vector representing the binary variable.}
}
\value{
A data frame containing the computed performance measures.
}
\description{
The \code{continuous_x_binary} function computes summary statistics and
 performs a t-test for paired continuous and binary variables.
}
\details{
The function calculates the following summary statistics for each
 binary category in the \code{b} variable:
\itemize{
\item Mean - the mean value of the continuous variable within each category.
\item Median - the median value of the continuous variable within each category.
\item Standard Deviation (SD) - the standard deviation of the continuous
 variable within each category.
\item 25th Percentile (Pct25) - the 25th percentile value of the continuous
 variable within each category.
\item 75th Percentile (Pct75) - the 75th percentile value of the continuous
 variable within each category.
}

The function also performs a t-test to compare the means of the
 continuous variable between the two categories.
 The t-test results include:
\itemize{
\item \code{ttest_t}: The t-statistic value from the t-test.
\item \code{ttest_pval}: The p-value from the t-test.
}
}
\examples{
# Example data
a <- c(10, 12, 15, 8, 7)
b <- c(TRUE, FALSE, TRUE, FALSE, TRUE)

# Compute summary statistics and perform t-test for paired continuous and binary variables
result <- continuous_x_binary(a, b)
print(result)

}
