% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-backend.R
\name{continuous_x_category}
\alias{continuous_x_category}
\title{continuous_x_category Function}
\usage{
continuous_x_category(a, b)
}
\arguments{
\item{a}{A numeric vector representing the continuous variable.}

\item{b}{A character vector representing the categorical variable.}
}
\value{
A data frame containing the computed performance measures.
}
\description{
The \code{continuous_x_category} function computes summary statistics and
 performs an ANOVA test for paired continuous and categorical variables.
}
\details{
The function calculates the following summary statistics for each
 category in the \code{b} variable:
\itemize{
\item Mean - the mean value of the continuous variable within each category.
\item Median - the median value of the continuous variable within each category.
\item Standard Deviation (SD) - the standard deviation of the continuous
 variable within each category.
\item 25th Percentile (Pct25) - the 25th percentile value of the continuous
 variable within each category.
\item 75th Percentile (Pct75) - the 75th percentile value of the continuous
 variable within each category.
}

The function also performs an ANOVA test to compare the means of the
 continuous variable across different categories.
 The ANOVA test results include:
\itemize{
\item \code{anova_f}: The F-statistic value from the ANOVA test.
\item \code{anova_pval}: The p-value from the ANOVA test.
}
}
\examples{
# Example data
a <- c(10, 12, 15, 8, 7)
b <- c("A", "B", "B", "A", "B")

# Compute summary statistics and perform ANOVA test for paired continuous and categorical variables
result <- continuous_x_category(a, b)
print(result)

}
