% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_predictions.R
\name{correct_predictions}
\alias{correct_predictions}
\title{Expected Percentage of Correct Predictions}
\usage{
correct_predictions(model, ci = 0.95)
}
\arguments{
\item{model}{Model with binary outcome.}

\item{ci}{The level of the confidence interval.}
}
\value{
A data frame with three columns: the expected percentage of
  correct predictions and the related lower and upper confidence interval.
}
\description{
Expected percentage of correct predictions (ePCP) for models
  with binary outcome.
}
\details{
The ePCP ranges from 0 to 1, where values closer to 1 mean that
  the model predicts the outcome better than models with an eCPC closer to 0.
}
\examples{
data(mtcars)
m <- glm(formula = vs ~ hp + wt, family = binomial, data = mtcars)
correct_predictions(m)

}
\references{
Herron, M. (1999). Postestimation Uncertainty in Limited Dependent Variable Models. Political Analysis, 8, 83–98.
}
