% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-DataObject.R
\name{create_data_object}
\alias{create_data_object}
\title{Create DataObject}
\usage{
create_data_object(var)
}
\arguments{
\item{var}{The input data, which can be a numeric vector, logical vector,
or character vector.}
}
\value{
An instance of the DataObject class containing the raw data,
 processed data, variable type, and a logical vector indicating the index
 of missing values in the data.
}
\description{
Creates a new instance of the DataObject class based on the input data.
}
\details{
The `create_data_object` function creates a new instance of the
 DataObject class based on the input data. It checks the type of the data
 using the \code{check_variable_type} function, and then processes the data
 accordingly. For binary data, missing values are set to 0. For category
 data, missing values are set to a new category "NA". For continuous data,
 missing values are set to the median of the non-missing values.
}
\examples{
# Create a DataObject instance
data_obj <- create_data_object(c(1, 2, NA, 4, 5))

# Access slots
raw_data <- data_obj@raw_data
processed_data <- data_obj@processed_data
variable_type <- data_obj@variable_type
na_index <- data_obj@na_index

}
