% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-DataObject.R
\docType{class}
\name{DataObject-class}
\alias{DataObject-class}
\title{DataObject Class}
\description{
A class to store data along with its processing information and
 variable type.
}
\details{
The `DataObject` class is designed to store data along with its
 processing information and variable type. The slot \code{raw_data} can hold
 any type of data, and \code{processed_data} stores the data after
 processing. The variable type is described by the \code{variable_type} slot,
 which can be one of "binary", "category", or "continuous". The
 \code{na_index} slot contains a logical vector indicating the index of
 missing values in the data.
}
\section{Slots}{

\describe{
\item{\code{raw_data}}{An object of class "ANY" containing the raw data.}

\item{\code{processed_data}}{An object of class "ANY" containing the processed data.}

\item{\code{variable_type}}{A character string indicating the variable type.}

\item{\code{na_index}}{A logical vector indicating the index of missing values
in the data.}
}}

