% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_nakagawa.R
\name{.badlink}
\alias{.badlink}
\title{helper-function, telling user if model is supported or not}
\usage{
.badlink(link, family)
}
\description{
helper-function, telling user if model is supported or not
}
\keyword{internal}
