% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_nakagawa.R
\name{.collapse_cond}
\alias{.collapse_cond}
\title{glmmTMB returns a list of model information, one for conditional and one for zero-inflated part, so here we "unlist" it}
\usage{
.collapse_cond(fit)
}
\description{
glmmTMB returns a list of model information, one for conditional and one for zero-inflated part, so here we "unlist" it
}
\keyword{internal}
