% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variance.R
\name{.get_variance_beta}
\alias{.get_variance_beta}
\title{Get distributional variance for beta-family}
\usage{
.get_variance_beta(mu, phi)
}
\description{
Get distributional variance for beta-family
}
\keyword{internal}
