% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variance.R
\name{.get_variance_dispersion}
\alias{.get_variance_dispersion}
\title{Get dispersion-specific variance}
\usage{
.get_variance_dispersion(x, vals, faminfo, obs.terms)
}
\description{
Get dispersion-specific variance
}
\keyword{internal}
