% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variance.R
\name{.get_variance_residual}
\alias{.get_variance_residual}
\title{Get residual (distribution specific) variance from random effects}
\usage{
.get_variance_residual(x, var.cor, faminfo, type)
}
\description{
Get residual (distribution specific) variance from random effects
}
\keyword{internal}
