% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looic.R
\name{looic}
\alias{looic}
\title{LOO-related Indices for Bayesian regressions.}
\usage{
looic(model)
}
\arguments{
\item{model}{A Bayesian regression model.}
}
\description{
Compute LOOIC and ELPD for Bayesian regressions.
}
\examples{
\dontrun{
library(rstanarm)

model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
looic(model)
}

}
