% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdonalds_omega.r
\name{mcdonalds_omega}
\alias{mcdonalds_omega}
\alias{mcdonalds_omega.data.frame}
\title{McDonald's Omega for Items or Scales}
\usage{
mcdonalds_omega(x, ...)

\method{mcdonalds_omega}{data.frame}(x, ci = 0.95, verbose = TRUE, ...)
}
\arguments{
\item{x}{A matrix or a data frame, or an object returned by
\verb{[parameters::principal_components()]}.}

\item{...}{Currently not used.}

\item{ci}{Confidence interval for the reliability estimate. If \code{NULL},
no confidence interval is computed.}
}
\value{
The McDonald's Omega value for \code{x}.
}
\description{
Compute various measures of internal consistencies
for tests or item-scales of questionnaires.
}
\details{
The McDonald's Omega value for \code{x}. A value closer to 1
indicates greater internal consistency, where usually following
rule of thumb is applied to interpret the results:
\ifelse{html}{\out{&omega;}}{\eqn{\omega}{omega}} < 0.5 is unacceptable,
0.5 < \ifelse{html}{\out{&omega;}}{\eqn{\omega}{omega}} < 0.6 is poor,
0.6 < \ifelse{html}{\out{&omega;}}{\eqn{\omega}{omega}} < 0.7 is questionable,
0.7 < \ifelse{html}{\out{&omega;}}{\eqn{\omega}{omega}} < 0.8 is acceptable,
and everything > 0.8 is good or excellent.

\code{mcdonalds_omega()} is a simplified implementation of the \code{MBESS::ci.reliability()}
function. Currently, it only computes the simple McDonald's Omega estimate
(not hierarchical, not for categorical data) and should return the same
results as the default \code{MBESS::ci.reliability()} call.
}
\note{
The code is based on the \code{MBESS::ci.reliability()} function, which
is licensed under the GPL-2|GPL-3 license. Credits go to Sunthud Pornprasertmanit
and Ken Kelley.
}
\examples{
\dontshow{if (requireNamespace("lavaan", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(iris)
x <- iris[1:4]
mcdonalds_omega(x)
\dontshow{\}) # examplesIf}
}
\references{
McDonald, R.P. (1999). Test theory: A unified treatment. Hillsdale: Erlbaum.
}
