% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.psych.R
\name{model_performance.fa}
\alias{model_performance.fa}
\title{Performance of FA / PCA models}
\usage{
\method{model_performance}{fa}(model, metrics = "all", verbose = TRUE, ...)
}
\arguments{
\item{model}{A model object of class \code{fa} (e.g., from \code{psych::fa()}),
\code{principal} (e.g., from \code{psych::principal()}), or from
\code{parameters::factor_analysis()}.}

\item{metrics}{Can be \code{"all"} or a character vector of metrics to be computed
(some of \code{"Chi2"}, \code{"Chi2_df"}, \code{"p_Chi2"}, \code{"RMSA"}, \code{"RMSA_corrected"},
\code{"TLI"}, \code{"RMSEA"}, \code{"RMSEA_CI_low"}, \code{"RMSEA_CI_high"}, and \code{"BIC"}.}

\item{verbose}{Toggle off warnings.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame (with one row) and one column per "index" (see
\code{metrics}).
}
\description{
Compute indices of model performance for models from the \strong{psych} package,
and for \code{parameters::factor_analysis()}.
}
\examples{
\dontshow{if (all(insight::check_if_installed(c("psych", "GPArotation", "psychTools"), quietly = TRUE))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
out <- psych::fa(psychTools::bfi[, 1:25], 5)
model_performance(out)
\dontshow{\}) # examplesIf}
}
