% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.glm.R
\name{model_performance.glm}
\alias{model_performance.glm}
\title{Performance of General Linear Models}
\usage{
\method{model_performance}{glm}(model, metrics = "all", ...)
}
\arguments{
\item{model}{Object of class \link{glm}.}

\item{metrics}{Can be \code{"all"} or a list of metrics to be computed (some of \code{c("AIC", "BIC", "R2")}).}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute indices of model performance for general linear models.
}
\examples{
model <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
model_performance(model)
}
