% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.lavaan.R
\name{model_performance.lavaan}
\alias{model_performance.lavaan}
\title{Performance of lavaan SEM / CFA Models}
\usage{
\method{model_performance}{lavaan}(model, metrics = "all", ...)
}
\arguments{
\item{model}{A \code{lavaan} model.}

\item{metrics}{Can be \code{"all"} or a character vector of metrics to be computed (some of \code{c("Chisq", "Chisq_DoF", "Chisq_p", "Baseline", "Baseline_DoF", "Baseline_p", "CFI", "TLI", "NNFI", "RFI", "NFI", "PNFI", "IFI", "RNI", "Loglikelihood", "AIC", "BIC", "BIC_adjusted", "RMSEA", "SMRM")}).}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame (with one row) and one column per "index" (see \code{metrics}).
}
\description{
Compute indices of model performance for SEM or CFA models from the \code{lavaan} package.
}
\details{
See the documentation for \code{lavaan::fitmeasures}.
}
\examples{
\dontrun{
library(lavaan)

# Confirmatory Factor Analysis (CFA) ---------

structure <- ' visual  =~ x1 + x2 + x3
               textual =~ x4 + x5 + x6
               speed   =~ x7 + x8 + x9 '
model <- lavaan::cfa(structure, data=HolzingerSwineford1939)
model_performance(model)
}

}
