% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model.R
\name{null_model}
\alias{null_model}
\title{Null-model}
\usage{
null_model(model)
}
\arguments{
\item{model}{A Statistical model.}
}
\description{
Generate a null-model (intercept and random effects only, no fixed effects).
}
