% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion.R
\name{overdispersion}
\alias{overdispersion}
\title{Check overdispersion of GL(M)M's}
\usage{
overdispersion(x, ...)
}
\arguments{
\item{x}{Fitted model of class \code{merMod}, \code{glmmTMB}, \code{glm},
or \code{glm.nb} (package \pkg{MASS}).}

\item{...}{Currently not used.}
}
\value{
Results from the on the overdispersion test, like chi-squared statistics,
   p-value or dispersion ratio.
}
\description{
\code{overdispersion()} checks generalized linear (mixed) models
   for overdispersion.
}
\details{
A p-value < .05 indicates overdispersion. For \code{merMod}- and
   \code{glmmTMB}-objects, \code{overdispersion()} is based on the code in
   the \href{http://bbolker.github.io/mixedmodels-misc/glmmFAQ.html}{GLMM FAQ},
   section \emph{How can I deal with overdispersion in GLMMs?}. Note that
   this function only returns an \emph{approximate} estimate of an
   overdispersion parameter, and is probably inaccurate for zero-inflated
   mixed models (fitted with \code{glmmTMB}). The same code is also used to
   check overdispersion for negative binomial models.
   \cr \cr
   For Poisson-models, the overdispersion test is based on the code
   from \cite{Gelman and Hill (2007), page 115}.
}
\examples{

}
\references{
Bolker B et al. (2017): \href{http://bbolker.github.io/mixedmodels-misc/glmmFAQ.html}{GLMM FAQ.}
 \cr \cr
 Gelman A, Hill J (2007) Data Analysis Using Regression and Multilevel/Hierarchical Models. Cambridge, New York: Cambridge University Press
}
