% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Summary of principal component analysis}
\usage{
pca(x)
}
\arguments{
\item{x}{A data frame or a \code{\link[stats]{prcomp}}-object.}
}
\value{
A data frame with all loadings of principal components.
}
\description{
This function performs a principal component analysis and
  returns the loadings as data frame.
}
\examples{
data(iris)
pca(iris[, 1:4])

}
