% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_rotate.R
\name{pca_rotate}
\alias{pca_rotate}
\title{Rotate loading matrix of principal component analysis}
\usage{
pca_rotate(x, n_comp = NULL, rotation = c("varimax", "quartimax",
  "promax", "oblimin", "simplimax", "cluster", "none"))
}
\arguments{
\item{x}{A data frame or a \code{\link[stats]{prcomp}} object.}

\item{n_comp}{Number of components to extract. If \code{rotation = "varmiax"}
and \code{n_comp = NULL}, number of components is based on the Kaiser-criteria.}

\item{rotation}{Rotation of the factor loadings. May be one of
\code{"varimax", "quartimax", "promax", "oblimin", "simplimax", "cluster"}
or \code{"none"}.}
}
\value{
A rotated loadings matrix, as data frame. Details on the variance
  components are saved as attributes.
}
\description{
...
}
\details{
The \code{print()}-method for \code{pca_rotate()} has a
   \code{cutoff}-argument, which is a scalar between 0 and 1, indicating
   which (absolute) values from the loadings should be blank in the
   output. By default, all loadings between -.1 and .1 are not shown.
}
\examples{
data(iris)
pca_rotate(iris[, 1:4], n_comp = 2)

pr <- pca_rotate(iris[, 1:4], n_comp = 2)

# show all
print(pr, cutoff = .001)

# show only some
print(pr, cutoff = .5)

}
