% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-frontend.R
\name{performance}
\alias{performance}
\title{Performance Measures for Pairs of Variables}
\usage{
performance(yhat, y, use_complete_obs = FALSE)
}
\arguments{
\item{yhat}{A vector representing the first variable to be compared.}

\item{y}{A vector representing the second variable to be compared.}

\item{use_complete_obs}{Logical. If \code{TRUE}, the function will use only
pairwise complete observations. Default is \code{FALSE}.}
}
\value{
A data frame containing the computed performance measures for the
 given pair of variables.
}
\description{
The \code{performance} function allows users to analyze the performance
 measures for pairs of variables. It takes two inputs \code{var1} and
 \code{var2}, representing the two variables to be compared.
 Users can specify whether to use only pairwise complete observations using
 the \code{use_complete_obs} argument (default is \code{FALSE}).
 The function utilizes various sub-functions to compute the appropriate
 performance measures based on the variable types of
 \code{var1} and \code{var2}.
}
\examples{
# Example usage
var1 <- c(1, 2, 3, NA, 5)
var2 <- c(3, 2, 1, 1, 0)

# Compute performance measures for continuous and binary variables
result <- performance(var1, var2, use_complete_obs = TRUE)

}
