% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_LOO.R
\name{performance_LOO}
\alias{performance_LOO}
\title{LOO-related Indices for Bayesian regressions.}
\usage{
performance_LOO(model)
}
\arguments{
\item{model}{A Bayesian regression model.}
}
\description{
Compute LOOIC and ELPD for Bayesian regressions.
}
\examples{
\dontrun{
library(rstanarm)

model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
performance_LOO(model)
}


}
