% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_R2_bayesian.R
\name{performance_r2_bayesian}
\alias{performance_r2_bayesian}
\title{R2 for Bayesian regressions.}
\usage{
performance_r2_bayesian(model)
}
\arguments{
\item{model}{A Bayesian regression model.}
}
\description{
Compute R2 for Bayesian models For mixed models (including a random part), it additionally computes the R2 related to the fixed effects only.
}
\examples{
\dontrun{
library(rstanarm)

model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
performance_r2_bayesian(model)

model <- rstanarm::stan_lmer(Petal.Length ~ Petal.Width + (1|Species), data = iris)
performance_r2_bayesian(model)
}

}
\references{
Gelman, A., Goodrich, B., Gabry, J., & Vehtari, A. (2018). R-squared for Bayesian regression models. The American Statistician, The American Statistician, 1-6.
}
