% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_r2_LOO_adjusted.R
\name{performance_r2_LOO_adjusted}
\alias{performance_r2_LOO_adjusted}
\title{LOO-adjusted R2.}
\usage{
performance_r2_LOO_adjusted(model)
}
\arguments{
\item{model}{A Bayesian regression model.}
}
\description{
Compute LOO-adjusted R2.
}
\examples{
\dontrun{
library(rstanarm)

model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
performance_r2_LOO_adjusted(model)
}


}
