% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_R2_Tjur.R
\name{performance_R2_tjur}
\alias{performance_R2_tjur}
\alias{cod}
\title{Tjur's (2009) coefficient of determination (D).}
\usage{
performance_R2_tjur(model)

cod(model)
}
\arguments{
\item{model}{Binomial Model.}
}
\description{
Computes Tjur's (2009) coefficient of determination.
}
\note{
This method calculates the Coefficient of Discrimination \code{D}
   for generalized linear (mixed) models for binary outcomes. It is an
   alternative to other Pseudo-R-squared values like Nagelkerke's
   R2 or Cox-Snell R2. The Coefficient of Discrimination \code{D}
   can be read like any other (Pseudo-)R-squared value.
}
\examples{
model <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
performance_R2_tjur(model)

\dontrun{
library(rstanarm)
model <- rstanarm::stan_glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
}


}
\references{
Tjur, T. (2009). Coefficients of determination in logistic regression models—A new proposal: The coefficient of discrimination. The American Statistician, 63(4), 366-372.
}
