% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_reliability.R
\name{performance_reliability}
\alias{performance_reliability}
\alias{performance_dvour}
\title{Random Effects Reliability}
\usage{
performance_reliability(x, ...)

performance_dvour(x, ...)
}
\arguments{
\item{x}{A model object (or from \code{\link[modelbased:estimate_grouplevel]{modelbased::estimate_grouplevel()}}).}

\item{...}{Currently not used.}
}
\description{
TODO: Add description.
}
\details{
Variability-Over-Uncertainty Ratio (d-vour).

TODO: Add details.

Intepretation: d-vour corresponds to: between-groups variability /  (between-groups variability + within-group variability)
A d-vour of 3/4 (0.75) means that there is 3 times more variability between groups than within groups (3/1),
and a d-vour of less than 0.5 means that there is more variability within groups than between groups (which is
bad if the goal is to analyze group-level effects).
}
\examples{
\dontshow{if (require("lme4") && require("glmmTMB")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df <- read.csv("https://raw.githubusercontent.com/easystats/circus/refs/heads/main/data/illusiongame.csv")

m <- lme4::lmer(RT ~ (1 | Participant), data = df)
performance_reliability(m)
performance_dvour(m)
m <- glmmTMB::glmmTMB(RT ~ (1 | Participant), data = df)
performance_reliability(m)
performance_dvour(m)

m <- lme4::lmer(RT ~ (1 | Participant) + (1 | Trial), data = df)
performance_reliability(m)
performance_dvour(m)
m <- glmmTMB::glmmTMB(RT ~ (1 | Participant) + (1 | Trial), data = df)
performance_reliability(m)
performance_dvour(m)

m <- lme4::lmer(RT ~ Illusion_Difference + (Illusion_Difference | Participant) + (1 | Trial), data = df)
performance_reliability(m)
performance_dvour(m)
m <- glmmTMB::glmmTMB(RT ~ Illusion_Difference + (Illusion_Difference | Participant) + (1 | Trial), data = df)
performance_reliability(m)
performance_dvour(m)
\dontshow{\}) # examplesIf}
}
\references{
TODO.
}
