% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_roc.R
\name{performance_roc}
\alias{performance_roc}
\alias{performance_roc.numeric}
\alias{performance_roc.glm}
\title{Simple ROC curve}
\usage{
performance_roc(x, ...)

\method{performance_roc}{numeric}(x, predictions, ...)

\method{performance_roc}{glm}(x, new_data = NULL, ...)
}
\arguments{
\item{x}{A numeric vector, representing the outcome (0/1), or a model with
binomial outcome.}

\item{predictions}{If \code{x} is numeric, a numeric vector of same length
as \code{x}, representing the actual predicted values.}

\item{new_data}{If \code{x} is a model, a data frame that is passed to
\code{predict()} as \code{newdata}-argument.}
}
\value{
A data frame with two columns, the x/y-coordinate pairs for the ROC
curve (\code{Sensivity} and \code{Specifity}), which, for instance, can be
used to further compute the AUC.
}
\description{
This function calculates a simple ROC curves of x/y coordinates
based on response and predictions of a binomial model.
}
\examples{
library(bayestestR)
data(iris)
set.seed(123)
iris$y <- rbinom(nrow(iris), size = 1, .3)
folds <- sample(nrow(iris), size = nrow(iris) / 8, replace = FALSE)
test_data <- iris[folds, ]
train_data <- iris[-folds, ]

model <- glm(y ~ Sepal.Length + Sepal.Width, data = train_data, family = "binomial")
performance_roc(model, test_data)

roc <- performance_roc(model, test_data)
area_under_curve(roc$Sensivity, roc$Specifity)

}
