% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check.R
\name{pp_check}
\alias{pp_check}
\title{Posterior predictive checks for frequentist models}
\usage{
pp_check(object, ...)
}
\arguments{
\item{object}{A statistical model.}

\item{...}{Not used.}

\item{iterations}{The number of draws to simulate/bootstrap.}
}
\value{
A data frame
}
\description{
Posterior predictive checks for frequentist models.
}
\details{
Posterior predictive checks means \dQuote{simulating replicated data
  under the fitted model and then comparing these to the observed data}
  \cite{(Gelman and Hill, 2007, p. 158)}. Posterior predictive checks
  can be used to \dQuote{look for systematic discrepancies between real and
  simulated data}  \cite{(Gelman et al. 2014, p. 169)}.
}
\examples{
model <- lm(Sepal.Length ~ Species * Petal.Width + Petal.Length, data = iris)
if (require("ggplot2") && require("see")) {
  pp_check(model)
}
}
\references{
\itemize{
  \item Gelman, A., & Hill, J. (2007). Data analysis using regression and multilevel/hierarchical models. Cambridge; New York: Cambridge University Press.
  \item Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari, A., & Rubin, D. B. (2014). Bayesian data analysis. (Third edition). CRC Press.
}
}
