% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_loo.R
\name{r2_loo}
\alias{r2_loo}
\title{LOO-adjusted R2}
\usage{
r2_loo(model)
}
\arguments{
\item{model}{A Bayesian regression model.}
}
\description{
Compute LOO-adjusted R2.
}
\examples{
\dontrun{
library(rstanarm)

model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
r2_loo(model)
}

}
