% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_mcfadden.R
\name{r2_mcfadden}
\alias{r2_mcfadden}
\title{McFadden's (1973) R2}
\usage{
r2_mcfadden(model)
}
\arguments{
\item{model}{Multinomial Logit (\code{mlogit}) Model.}
}
\description{
DESCRIPTION TO BE IMPROVED.
}
\references{
\itemize{
  \item McFadden, D. (1987). Regression-based specification tests for the multinomial logit model. Journal of econometrics, 34(1-2), 63-82.
  \item McFadden, D. (1973). Conditional logit analysis of qualitative choice behavior.
}
}
