% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_nakagawa.R
\name{r2_nakagawa}
\alias{r2_nakagawa}
\title{Nakagawa's (2013, 2018) R2 for mixed models}
\usage{
r2_nakagawa(model)
}
\arguments{
\item{model}{A statistical model.}
}
\description{
DESCRIPTION TO BE IMPROVED.
}
\examples{
\dontrun{
library(lme4)
model <- lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
r2_nakagawa(model)
}

}
