% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_residuals.R
\name{simulate_residuals}
\alias{simulate_residuals}
\title{Simulate randomized quantile residuals from a model}
\usage{
simulate_residuals(x, iterations = 250, ...)
}
\arguments{
\item{x}{A model object.}

\item{iterations}{Number of simulations to run.}

\item{...}{Arguments passed on to \code{\link[DHARMa:simulateResiduals]{DHARMa::simulateResiduals()}}.}
}
\value{
Simulated residuals.
}
\description{
to do.
}
\details{
Based on \code{\link[DHARMa:simulateResiduals]{DHARMa::simulateResiduals()}}. See also \code{vignette("DHARMa")}.
}
\examples{
\dontshow{if (require("DHARMa")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
m <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
simulate_residuals(m)
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item Hartig, F., & Lohse, L. (2022). DHARMa: Residual Diagnostics for Hierarchical
(Multi-Level / Mixed) Regression Models (Version 0.4.5). Retrieved from
https://CRAN.R-project.org/package=DHARMa
\item Dunn, P. K., & Smyth, G. K. (1996). Randomized Quantile Residuals. Journal
of Computational and Graphical Statistics, 5(3), 236. \doi{10.2307/1390802}
}
}
