% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backports.R
\name{str2lang}
\alias{str2lang}
\alias{isTRUE}
\alias{isFALSE}
\alias{deparse1}
\title{Backports of newer R functions}
\usage{
str2lang(s)

isTRUE(x)

isFALSE(x)

deparse1(expr, collapse = " ", width.cutoff = 500L, ...)
}
\description{
Backports of newer R functions.
}
\keyword{internal}
