% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wald.R
\name{test_wald}
\alias{test_wald}
\title{Test if Models are Different using Wald Test}
\usage{
test_wald(...)
}
\arguments{
\item{...}{Multiple model objects.}
}
\value{
A data frame.
}
\description{
The Wald test (also called the Wald Chi-Squared Test) is a rough approximation of the Likelihood Ratio Test (see \code{test_lrt()}). That said, it is more applicable than the LRT: you can often run a Wald test in situations where no other test can be run. Agresti (1990) suggests that you should use the LRT instead of the Wald test for small sample sizes or if the parameters are large. A "small" sample size is under about 30.
}
\examples{
# Nested Models
m1 <- lm(Sepal.Length ~ Petal.Width * Species, data = iris)
m2 <- lm(Sepal.Length ~ Petal.Width + Species, data = iris)
m3 <- lm(Sepal.Length ~ Petal.Width, data = iris)

test_wald(m1, m2, m3)  # equivalent to anova(m1, m2, m3)
}
\seealso{
\code{\link[=compare_performance]{compare_performance()}} to compare performance of many different models.
}
