% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_markdown.R, R/to_table.R
\name{table_to_markdown.performance_model}
\alias{table_to_markdown.performance_model}
\alias{to_table.performance_model}
\title{Export tables into different output formats}
\usage{
table_to_markdown.performance_model(x, digits = 2, ...)

\method{to_table}{performance_model}(x, format = "markdown", digits = 2, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=model_performance]{model_performance()}}
or \code{\link[=compare_performance]{compare_performance()}}.
or its summary.}

\item{digits}{Number of decimal places.}

\item{...}{Currently not used.}

\item{format}{String, indicating the output format. Currently, only
\code{"markdown"} is supported.}
}
\value{
A character vector. If \code{format = "markdown"}, the return value
  will be a character vector in markdown-table format.
}
\description{
Export tables (i.e. data frame) into different output formats.
  \code{table_to_markdown()} is a alias for \code{to_table(format = "markdown")}.
}
\details{
\code{to_table()} is useful when the table-output from functions,
  which is usually printed as formatted text-table to console, should
  be formatted for pretty table-rendering in markdown documents, or if
  knitted from rmarkdown to PDF or Word files.
}
\examples{
model <- lm(mpg ~ wt + cyl, data = mtcars)
mp <- model_performance(model)
to_table(mp)
}
