% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_collinearity.R
\name{check_collinearity}
\alias{check_collinearity}
\alias{check_collinearity.glmmTMB}
\title{Check for multicollinearity of model predictors}
\usage{
check_collinearity(x, ...)

\method{check_collinearity}{glmmTMB}(x, component = c("conditional",
  "count", "zi", "zero_inflated"), ...)
}
\arguments{
\item{x}{A model object (that should at least respond to \code{vcov()},
and if possible, also to \code{model.matrix()} - however, it also should
work without \code{model.matrix()}).}

\item{...}{Currently not used.}

\item{component}{For models with zero-inflation component, multicollinearity
can be checked for the conditional model (count component,
\code{component = "conditional"} or \code{component = "count"}) or
zero-inflation component (\code{component = "zero_inflated"} or
\code{component = "zi"}). Following model-classes are currently supported:
\code{hurdle}, \code{zeroinfl}, \code{zerocount}, \code{MixMod} and
\code{glmmTMB}.}
}
\value{
A data frame with three columns: The name of the model term, the
  variance inflation factor and the factor by which the standard error
  is increased due to possible correlation with other predictors.
}
\description{
\code{check_collinearity()} checks regression models for
  multicollinearity by calculating the variance inflation factor (VIF).
}
\details{
The variance inflation factor is a measure to analyze the magnitude
  of multicollinearity of model predictors. A VIF less than 5 indicates
  a low correlation of that predictor with other predictors. A value between
  5 and 10 indicates a moderate correlation, while VIF values larger than 10
  are a sign for high, not tolerable correlation of model predictors. The
  \emph{Increased SE} column in the output indicates how much larger
  the standard error is due to the correlation with other predictors.
}
\examples{
m <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
check_collinearity(m)

}
\references{
James, G., Witten, D., Hastie, T., & Tibshirani, R. (Hrsg.). (2013). An introduction to statistical learning: with applications in R. New York: Springer.
}
