% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dag.R
\name{check_dag}
\alias{check_dag}
\title{Check correct model adjustment for identifying causal effects}
\usage{
check_dag(
  ...,
  outcome = NULL,
  exposure = NULL,
  adjusted = NULL,
  effect = c("all", "total", "direct")
)
}
\description{
Check correct model adjustment for identifying causal effects.
}
