% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_group_variation.R
\name{check_group_variation}
\alias{check_group_variation}
\alias{check_group_variation.default}
\alias{check_group_variation.data.frame}
\title{Check variables for within- and/or between-group variation}
\usage{
check_group_variation(x, ...)

\method{check_group_variation}{default}(x, ...)

\method{check_group_variation}{data.frame}(
  x,
  select = NULL,
  by = NULL,
  include_by = FALSE,
  numeric_as_factor = FALSE,
  tolerance_numeric = 1e-04,
  tolerance_factor = "crossed",
  ...
)
}
\arguments{
\item{x}{A data frame or a mixed model. See details.}

\item{...}{Arguments passed to other methods}

\item{select}{Character vector (or formula) with names of variables to select
that should be checked. If \code{NULL}, selects all variables (except those in
\code{by}).}

\item{by}{Character vector (or formula) with the name of the variable that
indicates the group- or cluster-ID. For cross-classified or nested designs,
\code{by} can also identify two or more variables as group- or cluster-IDs.}

\item{include_by}{When there is more than one grouping variable, should they
be check against each other?}

\item{numeric_as_factor}{Should numeric variables be tested as factors?}

\item{tolerance_numeric}{The amount of variation (calculated by \code{var()}, i.e.
the variance of a variable) that is tolerated to indicate no within- or
between-effect.}

\item{tolerance_factor}{How should a non-numeric variable be identified as
varying only "within" a grouping variable? Options are:
\itemize{
\item \code{"crossed"} - if all groups have all unique values of X.
\item \code{"balanced"} - if all groups have all unique values of X, \emph{with equal
frequency}.
}}
}
\value{
A data frame with group, variable, and type columns.
}
\description{
\code{check_group_variation()} checks if variables a within- and/or between-effect,
i.e. if they vary within or between certain groups.
}
\details{
This function attempt to identify the hierarchical design of a dataset with
respect to grouping variables (\code{by}).
\cr\cr
If \code{x} is a (mixed effect) model, the variability of the fixed effects
predictors are checked with respect to the random grouping variables.
\subsection{Numeric variables}{

Numeric variables are portioned via \code{\link[datawizard:demean]{datawizard::demean()}} to their within-
and between-group components. Then, the variance for each variable's within-
and between-group component is calculated. Variable with within-group
variance larger than \code{tolerance_numeric} are labeled as \emph{within}, variable with
between-group variance larger than \code{tolerance_numeric} are labeled as \emph{between},
and variables with both variances larger than \code{tolerance_numeric} are labeled as
\emph{both}.

Setting \code{numeric_as_factor = TRUE} causes numeric variables to be tested
using the following criteria.
}

\subsection{Non-numeric variables}{

These variables can have one of the following 4 labels:
\itemize{
\item \emph{between} - the variable is fixed (has exactly one unique, constant value) for each group.
\item \emph{nested} - the variable varies within each group, with each group having their own set of
unique levels of the variable.
\item \emph{within} - the variable is \emph{crossed} with the grouping variable - each value appear
within each group. The \code{tolerance_factor} argument controls if full balance is also required.
\item \emph{both} - the variable is partially nested within the grouping variable (or, when
\code{tolerance_factor = "balanced"} the variable is fully crossed, but not perfectly balanced).
}
}
}
\examples{
set.seed(123)
dat <- data.frame(
  id = rep(letters, each = 2),
  between = rep(rnorm(26), each = 2),
  within = rep(rnorm(2), times = 26),
  both = rnorm(52)
)
check_group_variation(dat, by = "id")

}
