% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_group_variation.R
\name{check_group_variation}
\alias{check_group_variation}
\title{Check variables for within- and/or between-group variation}
\usage{
check_group_variation(
  x,
  select = NULL,
  by = NULL,
  include_by = FALSE,
  numeric_as_factor = FALSE,
  num_tolerance = 1e-04,
  fct_tolerance = "crossed"
)
}
\arguments{
\item{x}{A data frame.}

\item{select}{Character vector (or formula) with names of variables to select
that should be checked. If \code{NULL}, selects all variables (except those in
\code{by}).}

\item{by}{Character vector (or formula) with the name of the variable that
indicates the group- or cluster-ID. For cross-classified or nested designs,
\code{by} can also identify two or more variables as group- or cluster-IDs.}

\item{include_by}{When there is more than one grouping variable, should they
be check against each other?}

\item{numeric_as_factor}{Should numeric variables be tested as factors?}

\item{num_tolerance}{The amount of variation (calculated by \code{var()}, i.e. the
variance of a variable) that is tolerated to indicate no within- or
between-effect.}

\item{fct_tolerance}{How should a non-numeric variable be identified as
varying only "within" a grouping variable? Options are:
\itemize{
\item \code{"crossed"} - if all groups have all unique values of X.
\item \code{"balanced"} - if all groups have all unique values of X, \emph{with equal frequency}.
}}
}
\value{
A data frame with group, variable, and type columns.
}
\description{
\code{check_group_variation()} checks if variables a within- and/or between-effect,
i.e. if they vary within or between certain groups.
}
\details{
This function attempt to identify the hierarchical design of a dataset with
respect to grouping variables (\code{by}).
\subsection{Numeric variables}{

Numeric variables are portioned via \code{\link[datawizard:demean]{datawizard::demean()}} to their within-
and between-group components. Then, the variance for each variable's within-
and between-group component is calculated. Variable with within-group
variance larger than \code{num_tolerance} are labeled as \emph{within}, variable with
between-group variance larger than \code{num_tolerance} are labeled as \emph{between},
and variables with both variances larger than \code{num_tolerance} are labeled as
\emph{both}.
\cr\cr
Setting \code{numeric_as_factor = TRUE} causes numeric variables to be tested
using the following criteria.
}

\subsection{Non-numeric variables}{

These variables can have one of the following 4 labels:
\itemize{
\item \emph{between} - the variable is fixed (has exactly one unique, constant value) for each group.
\item \emph{nested} - the variable varies within each group, with each group having their own set of
unique levels of the variable.
\item \emph{within} - the variable is \emph{crossed} with the grouping variable - each value appear
within each group. The \code{fct_tolerance} argument controls if full balance is also required.
\item \emph{both} - the variable is partially nested within the grouping variable (or, when
\code{fct_tolerance = "balanced"} the variable is fully crossed, but not perfectly balanced).
}
}
}
