% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.R
\name{check_model}
\alias{check_model}
\alias{check_model.default}
\title{Visual check of model assumptions}
\usage{
check_model(x, ...)

\method{check_model}{default}(
  x,
  dot_size = 2,
  line_size = 0.8,
  panel = TRUE,
  check = "all",
  alpha = 0.2,
  detrend = FALSE,
  ...
)
}
\arguments{
\item{x}{A model object.}

\item{...}{Currently not used.}

\item{dot_size}{Size of dot-geoms.}

\item{line_size}{Size of line-geoms.}

\item{panel}{Logical, if \code{TRUE}, plots are arranged as panels; else,
single plots for each diagnostic are returned.}

\item{check}{Character vector, indicating which checks for should be performed
and plotted. May be one or more of
\code{"all", "vif", "qq", "normality", "linearity", "ncv", "homogeneity", "outliers", "reqq"}.
\code{"reqq"} is a QQ-plot for random effects and only available for mixed models.
\code{"ncv"} is an alias for \code{"linearity"}, and checks for non-constant
variance, i.e. for heteroscedasticity, as well as the linear relationship.
By default, all possible checks are performed and plotted.}

\item{alpha}{The alpha level of the confidence bands. Scalar from 0 to 1.}

\item{detrend}{Should QQ/PP plots be detrended?}
}
\value{
The data frame that is used for plotting.
}
\description{
Visual check of model various assumptions (normality of residuals,
normality of random effects, linear relationship, homogeneity of variance,
multicollinearity).
}
\details{
The plot \strong{Linearity} checks the assumption of linear
relationship. However, the spread of dots also indicate possible
heteroscedasticity (i.e. non-constant variance); hence, the alias \code{"ncv"}
for this plot. \strong{Some caution is needed} when interpreting these plots.
Although these plots are helpful to check model assumptions, they do not
necessarily indicate so-called "lack of fit", e.g. missed non-linear
relationships or interactions. Thus, it is always recommended to also look
at \href{https://strengejacke.github.io/ggeffects/articles/introduction_partial_residuals.html}{effect plots, including partial residuals}.
}
\note{
This function just prepares the data for plotting. To create the plots,
\CRANpkg{see} needs to be installed. Furthermore, this function suppresses
all possible warnings. In case you observe suspicious plots, please refer to
the dedicated functions (like \code{check_collinearity()}, \code{check_normality()}
etc.) to get informative messages and warnings.
}
\examples{
\dontrun{
m <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
check_model(m)

if (require("lme4")) {
  m <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
  check_model(m, panel = FALSE)
}

if (require("rstanarm")) {
  m <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200)
  check_model(m)
}
}
}
