% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_reliability.R
\name{check_reliability}
\alias{check_reliability}
\title{Variability-Over-Uncertainty Ratio (D-vour) for Random Effects Reliability}
\usage{
check_reliability(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Currently not used.}
}
\description{
TODO: Add description.
}
\details{
TODO: Add details.
}
\examples{
\dontshow{if (require("lme4")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Add groups to the data
data <- iris
data$Place <- as.factor(rep(c("P1", "P2", "P3", "P4", "P5", "P6"), each = 25))

# lme4
m <- lme4::lmer(Sepal.Width ~ Petal.Width + (Petal.Width | Place), data = data)
check_reliability(m)
\dontshow{\}) # examplesIf}
}
\references{
TODO.
}
\seealso{
Other functions to check model assumptions and and assess model quality: 
\code{\link{check_autocorrelation}()},
\code{\link{check_collinearity}()},
\code{\link{check_convergence}()},
\code{\link{check_heteroscedasticity}()},
\code{\link{check_homogeneity}()},
\code{\link{check_model}()},
\code{\link{check_outliers}()},
\code{\link{check_overdispersion}()},
\code{\link{check_predictions}()},
\code{\link{check_singularity}()},
\code{\link{check_zeroinflation}()}
}
\concept{functions to check model assumptions and and assess model quality}
