% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_residuals.R
\name{check_residuals}
\alias{check_residuals}
\alias{check_residuals.default}
\title{Check uniformity of simulated residuals}
\usage{
check_residuals(x, ...)

\method{check_residuals}{default}(x, alternative = c("two.sided", "less", "greater"), ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=simulate_residuals]{simulate_residuals()}} or
\code{\link[DHARMa:simulateResiduals]{DHARMa::simulateResiduals()}}.}

\item{...}{Passed down to \code{\link[stats:ks.test]{stats::ks.test()}}.}

\item{alternative}{A character string specifying the alternative hypothesis.
See \code{\link[stats:ks.test]{stats::ks.test()}} for details.}
}
\value{
The p-value of the test statistics.
}
\description{
\code{check_residuals()} checks generalized linear (mixed) models for uniformity
of randomized quantile residuals, which can be used to identify typical model
misspecification problems, such as over/underdispersion, zero-inflation, and
residual spatial and temporal autocorrelation.
}
\details{
Uniformity of residuals is checked using a Kolmogorov-Smirnov test.
There is a \code{plot()} method to visualize the distribution of the residuals.
}
\examples{
\dontshow{if (require("DHARMa")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dat <- DHARMa::createData(sampleSize = 100, overdispersion = 0.5, family = poisson())
m <- glm(observedResponse ~ Environment1, family = poisson(), data = dat)
res <- simulate_residuals(m)
check_residuals(res)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=simulate_residuals]{simulate_residuals()}}
}
