% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_distribution.R
\docType{data}
\name{classify_distribution}
\alias{classify_distribution}
\title{Classify the distribution of a model-family using machine learning}
\format{
An object of class \code{function} of length 1.
}
\usage{
classify_distribution
}
\description{
Choosing the right distributional family for regression models is essential
to get more accurate estimates and standard errors. This function may help to
Machine learning model trained to classify distributions
}
\details{
Mean accuracy and Kappa of 0.86 and 0.85, repsectively.
}
\keyword{datasets}
