% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_distribution.R
\docType{data}
\name{classify_distribution}
\alias{classify_distribution}
\title{Machine learning model trained to classify distributions}
\format{An object of class \code{randomForest.formula} (inherits from \code{randomForest}) of length 19.}
\usage{
classify_distribution
}
\description{
Mean accuracy and Kappa of 0.86 and 0.85, repsectively.
}
\keyword{datasets}
