% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance_analysis.R
\name{dominance_analysis}
\alias{dominance_analysis}
\title{Dominance Analysis}
\usage{
dominance_analysis(model, ...)
}
\arguments{
\item{model}{An \code{insight}- and \code{performance::r2}-supported model.}

\item{...}{Arguments passed to \code{domir::domin} function.}
}
\value{
Object of class \code{dominance_analysis}.
}
\description{
Computes Dominance Statistics and Designations
}
\details{
Dominance analysis discussion here...
}
\examples{

  model <- glm(vs ~ cyl + carb + mpg, data = datasets::mtcars)

  r2(model)

  dominance_analysis(model)

}
\references{
\itemize{
\item Azen, R., & Budescu, D. V. (2003). The dominance analysis approach
for comparing predictors in multiple regression. Psychological Methods,
8(2), 129-148. doi:10.1037/1082-989X.8.2.129

\item Budescu, D. V. (1993). Dominance analysis: A new approach to the
problem of relative importance of predictors in multiple regression.
Psychological Bulletin, 114(3), 542-551. doi:10.1037/0033-2909.114.3.542

\item Groemping, U. (2007). Estimators of relative importance in linear
regression based on variance decomposition. The American Statistician,
61(2), 139-147. doi:10.1198/000313007X188252

\item Luchman, J. N., Lei, X., & Kaplan, S. A. (2020). Relative
Importance Analysis With Multivariate Models: Shifting the Focus from
Independent Variables to Parameter Estimates. Journal of Applied
Structural Equation Modeling, 4(2), 1-20. doi:10.47263/JASEM.4(2)02

\item Luchman, J. N. (2021). Determining relative importance in Stata
using dominance analysis: domin and domme. Stata Journal 21(2),
510-538. doi:10.1177/1536867X211025837
}
}
