% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance_analysis.R
\name{dominance_analysis}
\alias{dominance_analysis}
\title{Dominance Analysis}
\usage{
dominance_analysis(model, ...)
}
\arguments{
\item{model}{A model object supported by \code{performance::r2()}. See 'Details'.}

\item{...}{Not used at current.}
}
\value{
Object of class \code{"dominance_analysis"}.

An object of class \code{"dominance_analysis"} is a list composed of the
following elements:
\describe{
\item{\code{general_dominance}}{Vector of general dominance statistics.}
\item{\code{standardized}}{Vector of general dominance statistics normalized
to sum to 1.}
\item{\code{ranks}}{Vector of ranks applied to the general dominance
statistics.}
\item{\code{conditional_dominance}}{Matrix of conditional dominance
statistics.  Each row represents a predictor; each column represents an
the average increment to R2 with a specific number of predictors in
the model.}
\item{\code{complete_dominance}}{Logical matrix of complete dominance
designations. The predictors represented in each row are
cross-referenced with predictors in each column.  Whether the predictor
in each column dominates the predictor in each row is represented in the
logical value in the matrix.}
\item{\code{model_R2}}{Value of R2 value returned by the \code{r2()} method for
the model.}
}
}
\description{
Computes Dominance Analysis Statistics and Designations
}
\details{
Computes two decompositions of the model's R2 and returns
a matrix of designations from which predictor relative importance
determinations can be obtained.

The input model is parsed using \code{insight::find_predictors()}, does not
yet support interactions, transformations, or offsets applied in the
R formula, and will fail with an error if any such terms are detected.

The model submitted must accept an formula object as a \code{formula}
argument.  In addition, the model object must accept the data on which
the model is estimated as a \code{data} argument.  Formulas submitted
using object references (i.e., \code{lm(mtcars$mpg ~ mtcars$vs)}) and
functions that accept data as a non-\code{data} argument
(e.g., \code{survey::svyglm()} uses \code{design}) will fail.

Models that return \code{TRUE} for the \code{insight::model_info()}
function's values "is_bayesian", "is_mixed", "is_gam",
is_multivariate", "is_zero_inflated",
or "is_hurdle" are not supported at current.

When \code{performance::r2()} returns multiple values, only the first is used
by default.

The underlying \code{domir::domin()} function that implements the dominance
statistic and designation computations has only been tested to R version
3.5 and will fail with an error if called in R versions < 3.5.
}
\examples{
if (getRversion() >= "3.5.0" && require("domir")) {
  data(mtcars)
  model <- glm(vs ~ cyl + carb + mpg, data = mtcars, family = binomial())

  r2(model)
  dominance_analysis(model)
}
}
\references{
\itemize{
\item Azen, R., & Budescu, D. V. (2003). The dominance analysis approach
for comparing predictors in multiple regression. Psychological Methods,
8(2), 129-148. doi:10.1037/1082-989X.8.2.129
\item Budescu, D. V. (1993). Dominance analysis: A new approach to the
problem of relative importance of predictors in multiple regression.
Psychological Bulletin, 114(3), 542-551. doi:10.1037/0033-2909.114.3.542
\item Groemping, U. (2007). Estimators of relative importance in linear
regression based on variance decomposition. The American Statistician,
61(2), 139-147. doi:10.1198/000313007X188252
\item Luchman, J. N., Lei, X., & Kaplan, S. A. (2020). Relative
Importance Analysis With Multivariate Models: Shifting the Focus from
Independent Variables to Parameter Estimates. Journal of Applied
Structural Equation Modeling, 4(2), 1-20. doi:10.47263/JASEM.4(2)02
\item Luchman, J. N. (2021). Determining relative importance in Stata
using dominance analysis: domin and domme. Stata Journal 21(2),
510-538. doi:10.1177/1536867X211025837
}
}
\seealso{
\code{\link[domir:domin]{domir::domin()}}
}
\author{
Joseph Luchman
}
