% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{The Intraclass Correlation Coefficient (ICC) for mixed models}
\usage{
icc(model)
}
\arguments{
\item{model}{A statistical model.}
}
\description{
DESCRIPTION TO BE IMPROVED.
}
\examples{
\dontrun{
library(lme4)
model <- lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
icc(model)
}

}
