% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\alias{icc.brmsfit}
\title{The Intraclass Correlation Coefficient (ICC) for mixed models}
\usage{
icc(model, ...)

\method{icc}{brmsfit}(model, re.form = NULL, robust = TRUE,
  ci = 0.95, ...)
}
\arguments{
\item{model}{A mixed effects model of class \code{merMod}, \code{glmmTMB},
\code{MixMod}, \code{lme}, \code{mixed}, \code{stanreg} or \code{brmsfit}.}

\item{...}{Currently not used.}

\item{re.form}{Formula containing group-level effects to be considered in
the prediction. If \code{NULL} (default), include all group-level effects.
Else, for instance for nested models, name a specific group-level effect
to calculate the variance decomposition for this group-level.}

\item{robust}{Logical, if \code{TRUE}, the median instead of mean is used to
calculate the central tendency of the variances.}

\item{ci}{The Credible Interval level.}
}
\description{
This function calculates the intraclass-correlation coefficient
 (icc) - sometimes also called \emph{variance partition coefficient}
 (vpc) - for mixed effects models. The ICC is calculated for \code{merMod}
 (\pkg{lme4}), \code{glmmTMB} (\pkg{glmmTMB}), \code{MixMod} (\pkg{GLMMadpative}),
 \code{lme} (\pkg{nlme}), \code{mixed} (\pkg{afex}), and \code{stanreg}
 (\pkg{rstanarm}) objects and can be interpreted as \dQuote{the proportion
 of the variance explained by the grouping structure in the population}
 \cite{(Hox 2002: 15)}. For models fitted with the \pkg{brms}-package,
 a variance decomposition based on the posterior predictive distribution
 is calculated (see 'Details').
}
\details{
\strong{Calculation of the ICC}
 \cr \cr
 The ICC is calculated by dividing the random effect variance,
 \ifelse{html}{\out{&sigma;<sup>2</sup><sub>i</sub>}}{\eqn{\sigma^2_i}}, by
 the total variance, i.e. the sum of the random effect variance and the
 residual variance, \ifelse{html}{\out{&sigma;<sup>2</sup><sub>&epsilon;</sub>}}{\eqn{\sigma^2_\epsilon}}.
 \cr \cr
 \strong{Adjusted and conditional ICC}
 \cr \cr
 \code{icc()} calculates an adjusted and conditional ICC, which take
 all sources of uncertainty (i.e. of \emph{all random effects}) into account. While
 the adjusted ICC only relates to the random effects, the conditional ICC
 also takes the fixed effects variances into account (see \cite{Nakagawa et al. 2017}).
 Typically, the \emph{adjusted} ICC is of interest when the analysis of random
 effects is of interest. \code{icc()} returns a meaningful ICC also for more
 complex random effects structures, like models with random slopes or nested
 design (more than two levels) and is applicable for models with other distributions
 than Gaussian. For more details on the computation of the variances, see
 \code{\link[insight]{get_variance}}.
 \cr \cr
 \strong{ICC for unconditional and conditional models}
 \cr \cr
 Usually, the ICC is calculated for the null model ("unconditional model").
 However, according to \cite{Raudenbush and Bryk (2002)} or
 \cite{Rabe-Hesketh and Skrondal (2012)} it is also feasible to compute the ICC
 for full models with covariates ("conditional models") and compare how
 much, e.g., a level-2 variable explains the portion of variation in the grouping
 structure (random intercept).
 \cr \cr
 \strong{ICC for specific group-levels }
 \cr \cr
 The proportion of variance for specific levels related to  each other
 (e.g., similarity of level-1-units within level-2-units or level-2-units
 within level-3-units) must be calculated manually. Use \code{\link[insight]{get_variance}}
 to get the random intercept variances (between-group-variances) and residual
 variance of the model, and calculate the ICC for the various level correlations.
 \cr \cr
 For example, for the ICC between level 1 and 2: \cr
 \code{sum(insight::get_variance_intercept(model)) /} \cr
 \code{  (sum(insight::get_variance_intercept(model)) + insight::get_variance_residual(model))}
 \cr \cr
 For for the ICC between level 2 and 3: \cr
 \code{insight::get_variance_intercept(model)[2] /} \cr
 \code{  sum(insight::get_variance_intercept(model))}
 \cr \cr
 \strong{ICC for brms-models}
 \cr \cr
 If \code{model} is of class \code{brmsfit}, \code{icc()} calculates a
 variance decomposition based on the posterior predictive distribution. In
 this case, first, the draws from the posterior predictive distribution
 \emph{not conditioned} on group-level terms (\code{posterior_predict(..., re.form = NA)})
 are calculated as well as draws from this distribution \emph{conditioned}
 on \emph{all random effects} (by default, unless specified else in \code{re.form})
 are taken. Then, second, the variances for each of these draws are calculated.
 The "ICC" is then the ratio between these two variances. This is the recommended
 way to analyse random-effect-variances for non-Gaussian models. It is then possible
 to compare variances accross models, also by specifying different group-level
 terms via the \code{re.form}-argument.
 \cr \cr
 Sometimes, when the variance of the posterior predictive distribution is
 very large, the variance ratio in the output makes no sense, e.g. because
 it is negative. In such cases, it might help to use \code{robust = TRUE}.
}
\examples{
\dontrun{
library(lme4)
model <- lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
icc(model)
}

}
\references{
\itemize{
 \item Hox J. 2002. Multilevel analysis: techniques and applications. Mahwah, NJ: Erlbaum
 \item Nakagawa S, Johnson P, Schielzeth H (2017) The coefficient of determination R2 and intra-class correlation coefficient from generalized linear mixed-effects models revisted and expanded. J. R. Soc. Interface 14. \doi{10.1098/rsif.2017.0213}
 \item Rabe-Hesketh S, Skrondal A. 2012. Multilevel and longitudinal modeling using Stata. 3rd ed. College Station, Tex: Stata Press Publication
 \item Raudenbush SW, Bryk AS. 2002. Hierarchical linear models: applications and data analysis methods. 2nd ed. Thousand Oaks: Sage Publications
 }
}
