% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_discrimination.R
\name{item_discrimination}
\alias{item_discrimination}
\alias{item_totalcor}
\title{Discrimination of Questionnaire Items (Corrected Item-Total Correlation)}
\usage{
item_discrimination(x, standardize = FALSE, verbose = TRUE)

item_totalcor(x, standardize = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{A matrix or a data frame.}

\item{standardize}{Logical, if \code{TRUE}, the data frame's vectors will be
standardized. Recommended when the variables have different measures /
scales.}

\item{verbose}{Toggle warnings and messages.}
}
\value{
A data frame with the item discrimination (\emph{corrected item-total
correlations}) for each item of the scale.
}
\description{
Compute various measures of internal consistencies for tests or
item-scales of questionnaires. \code{item_discrimination()} calculates the
corrected item-total correlations for each item of \code{x} with the remaining
items. \code{item_totalcor()} is an alias for \code{item_discrimination()}.
}
\details{
This function calculates the item discriminations (corrected item-total
correlations for each item of \code{x} with the remaining items) for each item
of a scale. The absolute value of the item discrimination indices should be
above \code{0.2}. An index between \code{0.2} and \code{0.4} is considered as "fair", while a
satisfactory index ranges from \code{0.4} to \code{0.7}. Items with low discrimination
indices are often ambiguously worded and should be examined. Items with
negative indices should be examined to determine why a negative value was
obtained (e.g. reversed answer categories regarding positive and negative
poles - in such cases, use \code{\link[datawizard:reverse]{datawizard::reverse()}} to reverse-code items
in advance).
}
\examples{
data(mtcars)
x <- mtcars[, c("cyl", "gear", "carb", "hp")]
item_discrimination(x)
}
\references{
\itemize{
\item Kelava A, Moosbrugger H (2020). Deskriptivstatistische Itemanalyse und
Testwertbestimmung. In: Moosbrugger H,  Kelava A, editors. Testtheorie und
Fragebogenkonstruktion. Berlin, Heidelberg: Springer, 143–158
}
}
