% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.merMod.R
\name{model_performance.merMod}
\alias{model_performance.merMod}
\title{Performance of Linear Models}
\usage{
\method{model_performance}{merMod}(model, metrics = "all", ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{metrics}{Can be \code{"all"} or a list of metrics to be computed (some of \code{c("AIC", "BIC", "R2", "R2_adj")}).}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute indices of model performance for linear models.
}
\examples{
model <- lm(mpg ~ wt + cyl, data = mtcars)
model_performance(model)
}
