% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{model_performance}
\alias{model_performance}
\title{Model Performance}
\usage{
model_performance(model, ...)
}
\arguments{
\item{model}{Statistical model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
See the documentation for your object's class:
\itemize{
 \item{\link[=model_performance.lm]{lm}}
 \item{\link[=model_performance.glm]{glm}}
 \item{\link[=model_performance.stanreg]{Bayesian lm}}
 }
}
