% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.bayesian.R
\name{model_performance.stanreg}
\alias{model_performance.stanreg}
\title{Performance of Bayesian Models}
\usage{
\method{model_performance}{stanreg}(model, metrics = "all", ci = 0.9,
  ...)
}
\arguments{
\item{model}{Object of class \link{lm} or \link{glm}.}

\item{metrics}{Can be \code{"all"} or a list of metrics to be computed (some of \code{c("LOO", "R2", "R2_adj")}).}

\item{ci}{The Credible Interval level for R2.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute indices of model performance for (general) linear models.
}
\examples{
\dontrun{
model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
model_performance(model)
}

}
\references{
Gelman, A., Goodrich, B., Gabry, J., & Vehtari, A. (2018). R-squared for Bayesian regression models. The American Statistician, The American Statistician, 1-6.
}
\seealso{
\link{performance_r2_bayesian}
}
