% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_mae.R
\name{performance_mae}
\alias{performance_mae}
\alias{mae}
\title{Mean Absolute Error of Models}
\usage{
performance_mae(model, ...)

mae(model, ...)
}
\arguments{
\item{model}{A model.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
Numeric, the mean absolute error of \code{model}.
}
\description{
Compute mean absolute error of models.
}
\examples{
data(mtcars)
m <- lm(mpg ~ hp + gear, data = mtcars)
performance_mae(m)
}
