% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_mor.R
\name{performance_mor}
\alias{performance_mor}
\title{Median Odds Ratio}
\usage{
performance_mor(x)
}
\arguments{
\item{x}{A (logistic) multilevel model.}
}
\value{
A data frame with two columns, one with the group (cluster) names and one
with the median odds ratios.
}
\description{
A measure of cluster-level variation in multilevel logistic regression,
defined as the median odds ratio between two randomly chosen individuals from
different clusters with identical covariates, comparing the person at higher
risk to the person at lower risk.
}
\examples{
\dontshow{if (all(insight::check_if_installed(c("lme4", "datawizdard"), quietly = TRUE))) withAutoprint(\{ # examplesIf}
data(sleepstudy, package = "lme4")
sleepstudy$mygrp <- sample(1:5, size = 180, replace = TRUE)
sleepstudy$high_reaction <- as.factor(datawizard::categorize(sleepstudy$Reaction))

m <- lme4::glmer(
  high_reaction ~ Days + (1 | Subject),
  data = sleepstudy,
  family = "binomial"
)
performance_mor(m)

m <- lme4::glmer(
  high_reaction ~ Days + (1 | mygrp) + (1 | Subject),
  data = sleepstudy,
  family = "binomial"
)
performance_mor(m)
\dontshow{\}) # examplesIf}
}
\references{
Larsen, K., & Merlo, J. (2005). Appropriate assessment of neighborhood effects
on individual health: integrating random and fixed effects in multilevel
logistic regression. American Journal of Epidemiology, 161(1), 81-88.
}
