% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_mor.R
\name{performance_mor}
\alias{performance_mor}
\title{Median Odds Ratio}
\usage{
performance_mor(x)
}
\arguments{
\item{x}{A (logistic) multilevel model.}
}
\value{
A data frame with two columns, one with the group (cluster) names and one
with the median odds ratios.
}
\description{
A measure of cluster-level variation in multilevel logistic regression,
defined as the median odds ratio between two randomly chosen individuals from
different clusters with identical covariates, comparing the person at higher
risk to the person at lower risk.
}
\details{
The MOR is always greater than or equal to 1 and can be interpreted as follows:
\itemize{
\item \emph{MOR close to 1:} No Cluster Effect. There is (almost) no between-cluster
heterogeneity, meaning cluster membership plays no role in the outcome.
\item \emph{MOR > 1:} Presence of Heterogeneity. Indicates meaningful variation across
clusters. If two persons are randomly picked from two different clusters,
the MOR represents the median factor by which the odds of the outcome
increase for the individual from the higher-risk cluster compared to the
individual in the lower-risk cluster.
}
}
\examples{
\dontshow{if (all(insight::check_if_installed(c("lme4", "datawizard"), quietly = TRUE))) withAutoprint(\{ # examplesIf}
data(sleepstudy, package = "lme4")
sleepstudy$mygrp <- sample(1:5, size = 180, replace = TRUE)
sleepstudy$high_reaction <- as.factor(datawizard::categorize(sleepstudy$Reaction))

m <- lme4::glmer(
  high_reaction ~ Days + (1 | Subject),
  data = sleepstudy,
  family = "binomial"
)
performance_mor(m)

m <- suppressWarnings(lme4::glmer(
  high_reaction ~ Days + (1 | mygrp) + (1 | Subject),
  data = sleepstudy,
  family = "binomial"
))
performance_mor(m)
\dontshow{\}) # examplesIf}
}
\references{
Larsen, K., & Merlo, J. (2005). Appropriate assessment of neighborhood effects
on individual health: integrating random and fixed effects in multilevel
logistic regression. American Journal of Epidemiology, 161(1), 81-88.
}
