% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_pcp.R
\name{performance_pcp}
\alias{performance_pcp}
\title{Percentage of Correct Predictions}
\usage{
performance_pcp(model, ci = 0.95)
}
\arguments{
\item{model}{Model with binary outcome.}

\item{ci}{The level of the confidence interval.}
}
\value{
A data frame with three columns: the percentage of
  correct predictions and the related lower and upper confidence interval.
}
\description{
Percentage of correct predictions (PCP) for models
  with binary outcome.
}
\details{
The PCP ranges from 0 to 1, where values closer to 1 mean that
  the model predicts the outcome better than models with an CPC closer to 0.
}
\examples{
data(mtcars)
m <- glm(formula = vs ~ hp + wt, family = binomial, data = mtcars)
performance_pcp(m)

}
\references{
Herron, M. (1999). Postestimation Uncertainty in Limited Dependent Variable Models. Political Analysis, 8, 83–98.
}
