% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_R2_bayesian.R
\name{performance_R2_bayesian}
\alias{performance_R2_bayesian}
\title{R2 for Bayesian regressions.}
\usage{
performance_R2_bayesian(model)
}
\arguments{
\item{model}{A Bayesian regression model.}
}
\description{
Compute R2 for Bayesian regressions.
}
\examples{
\dontrun{
library(rstanarm)

model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
performance_R2_bayesian(model)
}


}
