% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_R2_Tjur.R
\name{performance_R2_tjur}
\alias{performance_R2_tjur}
\title{Tjur's (2009) coefficient of determination (D).}
\usage{
performance_R2_tjur(model)
}
\arguments{
\item{model}{Binomial Model.}
}
\description{
Computes Tjur's (2009) coefficient of determination.
}
\examples{
model <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
performance_R2_tjur(model)

}
\references{
Tjur, T. (2009). Coefficients of determination in logistic regression models—A new proposal: The coefficient of discrimination. The American Statistician, 63(4), 366-372.
}
