% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{Compute the model's R2}
\usage{
r2(model, ...)
}
\arguments{
\item{model}{A statistical model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Returns a list containing values related to the most appropriate R2 for the given model. See the list below:
\itemize{
  \item logistic models (glm): \link[=r2_tjur]{Tjur's R2}
  \item General linear models (glm): \link[=r2_nagelkerke]{Nagelkerke's R2}
  \item Multinomial Logit (mlogit): \link[=r2_mcfadden]{McFadden's R2}
  \item Mixed models (merMod, glmmTMB): \link[=r2_nakagawa]{Nakagawa's R2}
  \item Bayesian models (brmsfit, stanreg): \link[=r2_bayes]{R2 bayes}
}
}
\examples{
model <- glm(vs ~ wt + mpg, data = mtcars, family = "binomial")
r2(model)

\dontrun{
library(lme4)
model <- lme4::lmer(Sepal.Length ~ Petal.Length + (1|Species), data=iris)
r2(model)
}

}
