% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\alias{r2.glm}
\alias{r2.lm}
\alias{r2.clm2}
\title{Compute r-squared}
\usage{
r2(x, ...)

\method{r2}{glm}(x, type = c("Nagelkerke", "CoxSnell"), ...)

\method{r2}{lm}(x, type = c("adjusted", "unadjusted"), ...)

\method{r2}{clm2}(x, type = c("Nagelkerke", "CoxSnell"), ...)
}
\arguments{
\item{x}{A fitted model.}

\item{...}{Currently not used.}

\item{type}{Character vector, indicating the type of r-squared that is
requested (like \code{"CoxSnell"} or \code{"adjusted"}).}
}
\value{
The requested r-squared value.
}
\description{
to do.
}
\details{
For linear models, the r-squared and adjusted r-squared value is returned,
         as provided by the \code{summary}-function.
         \cr \cr
         For generalized linear models, Cox & Snell's and Nagelkerke's
         pseudo r-squared values are returned.
}
\examples{
data(iris)

m <- lm(Sepal.Width ~ Species, data = iris)
r2(m)

}
