% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_nakagawa.R
\name{r2_nakagawa}
\alias{r2_nakagawa}
\title{Nakagawa's (2013, 2018) R2 for mixed models}
\usage{
r2_nakagawa(model, type = NULL, obj.name = NULL, fun.type = NULL)
}
\arguments{
\item{model}{A statistical model.}

\item{type}{TODO.}

\item{obj.name}{TODO.}

\item{fun.type}{TODO.}
}
\description{
DESCRIPTION TO BE IMPROVED.
}
