% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_zeroinflated.R
\name{r2_zeroinflated}
\alias{r2_zeroinflated}
\title{R2 for models with zero-inflation}
\usage{
r2_zeroinflated(model)
}
\arguments{
\item{model}{A model.}
}
\value{
The R2 value.
}
\description{
Calculates R2 for models with zero-inflation component, including mixed effects models.
}
\details{
\code{r2_zeroinflated()} is a correlation-based R2 measure, which
  is rather crude. It simply computes the squared correlation between the
  model's actual and predicted reponse.
}
\examples{
library(pscl)
data(bioChemists)
model <- zeroinfl(
  art ~ fem + mar + kid5 + ment | kid5 + phd,
  data = bioChemists
)

r2_zeroinflated(model)

}
