% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_markdown.R, R/to_table.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{table_to_markdown}
\alias{to_table}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{insight}{\code{\link[insight:to_table]{table_to_markdown}}, \code{\link[insight]{to_table}}}
}}

