% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_performance.R
\name{test_performance}
\alias{test_performance}
\title{Test if Models are Different}
\usage{
test_performance(..., reference = 1)
}
\arguments{
\item{...}{Multiple model objects.}

\item{reference}{Which model should be taken as a reference, against which all the other models are tested.}
}
\value{
A data frame.
}
\description{
This function aggregates different tests to compare different models. It tries to pick the appropriate based on whether the models have 'nested' predictors or 'non-nested' (i.e., different predictors).
}
\examples{
# Nested Models
m1 <- lm(Sepal.Length ~ Petal.Width * Species, data = iris)
m2 <- lm(Sepal.Length ~ Petal.Width + Species, data = iris)
m3 <- lm(Sepal.Length ~ Petal.Width, data = iris)

test_performance(m1, m2, m3)
test_performance(m1, m2, m3, include_formula=TRUE)

# Non-nested
if (require("lme4") & require("mgcv")) {
  m1 <- lm(Sepal.Length ~ Petal.Length + Species, data = iris)
  m2 <- lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
  m3 <- gam(Sepal.Length ~ s(Petal.Length, by = Species) + Species, data = iris)

  test_performance(m1, m2, m3)
}
}
\seealso{
\code{\link[=compare_performance]{compare_performance()}} to compare performance of many different models.
}
