% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_performance.R
\name{test_performance}
\alias{test_performance}
\title{Test if Models are Different}
\usage{
test_performance(...)
}
\arguments{
\item{...}{Multiple model objects.}
}
\value{
A data frame.
}
\description{
Test if models are different using different tests.
}
\examples{
# Nested Models
m1 <- lm(Sepal.Length ~ Petal.Width * Species, data=iris)
m2 <- lm(Sepal.Length ~ Petal.Width + Species, data=iris)
m3 <- lm(Sepal.Length ~ Petal.Width, data=iris)

test_performance(m1, m2, m3)

# Non-nested
if(require("lme4") & require("mgcv")){
m1 <- lm(Sepal.Length ~ Petal.Length + Species, data=iris)
m2 <- lme4::lmer(Sepal.Length ~ Petal.Length + (1|Species), data=iris)
m3 <- mgcv::gam(Sepal.Length ~ s(Petal.Length, by=Species) + Species, data=iris)

test_performance(m1, m2, m3)
}


}
\seealso{
\code{\link[=compare_performance]{compare_performance()}} to compare performance of many different models.
}
